refactor: update header file includes to use consistent casing for Server and Channel classes
