+++ include/Channel.hpp
