+++ src/commands/nick.cpp
@@ -1,4 +1,4 @@
-#include <server.hpp>
+#include <Server.hpp>
 #include <Client.hpp>
 #include <message.hpp>
 #include <responses.hpp>
