+++ src/server/Server.cpp
@@ -1,4 +1,4 @@
-#include <server.hpp>
+#include <Server.hpp>
 #include <common.hpp>
 #include <SocketManager.hpp>
 #include <EventLoop.hpp>
