+++ tests/tests.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include <server.hpp>
+#include <Server.hpp>
 #include <sys/socket.h>
 #include <netinet/in.h>
 #include <arpa/inet.h>
